smol fix for process_remove_member.rs (#1166)
* Update process_remove_member.rs

plz sers, I will earn so much of my hydra wasted efforts back if we can fix the remove Member ix.

expected behavior: feed it a member (which is seeded as mint of membership voucher) and a destination (native account) and it works
actual behavior: it does a check to see if member is owned by spl token, errors if so with "InvalidCloseAccountDestination. Error Number: 6024. Error Message: Sending Sol to a SPL token destination will render the sol unusable."
fix: this check is meant for the destination, to see if it is a native account. do a switcheroo and the checks should pass.

* Update process_remove_member.rs

it's actually better like this, tbf.

* Update error.rs

* too many :s

* solana program doesn't exist

* whitespace

* Fixing lint.

* Bump cli version.

---------

Co-authored-by: Jarett Dunn <jarettdunn@Jaretts-Laptop.local>
Co-authored-by: Blockiosaurus <blockiosaurus@gmail.com>
Co-authored-by: stranzhay <strange1120@gmail.com>